+++ pathos/sources/codesrc/utils/mxtk/include/mx/mxbutton.h
@@ -33,9 +33,7 @@ class mxButton : public mxWidget
 
 public:
 	// CREATORS
-
-	mxButton (mxWindow *parent, int x, int y, int w, int h, const char *label = 0, int id = 0);
-
+	mxButton(mxWindow* parent, int x, int y, int w, int h, const char* label = 0, __int64 id = 0);
 	virtual ~mxButton ();
 
 private:
