+++ pathos/sources/codesrc/utils/mxtk/include/mx/mxslider.h
@@ -40,7 +40,7 @@ public:
 	enum { Horizontal, Vertical };
 
 	// CREATORS
-	mxSlider (mxWindow *parent, int x, int y, int w, int h, int id = 0, int style = 0, bool ticks = false);
+	mxSlider(mxWindow* parent, int x, int y, int w, int h, __int64 id = 0, int style = 0, bool ticks = false);
 	virtual ~mxSlider ();
 
 	// MANIPULATORS
