+++ pathos/sources/codesrc/utils/mxtk/include/mx/mxtogglebutton.h
@@ -33,7 +33,7 @@ class mxToggleButton : public mxWidget
 
 public:
 	// CREATORS
-	mxToggleButton (mxWindow *parent, int x, int y, int w, int h, const char *label = 0, int id = 0);
+	mxToggleButton (mxWindow *parent, int x, int y, int w, int h, const char *label = 0, __int64 id = 0);
 	virtual ~mxToggleButton ();
 
 	// MANIPULATORS
